Update initialize-program-variables.fx application variables
Allow for application deadlines, registration fees, and deposits to be overridden.

Refactor to use those page variables instead of accessing |application| directly. The only time |application| is accessed is for form URLs and |allow-mail-registration|.

May consider converting this to a "private" front matter variable to avoid encouragement of using it directly.