T&C update and fix #28
Refactor cancellation policy for programs. Split the cancellation include into two parts so that we can list all the schedules for all the programs in the T&C page. If the URL to the T&C page includes a "program" query parameter, then only that program will appear on the page. This should be useful for linking back from the form applications.