Refactor excursions and fix #26
Call the collective group of excursions and activities as "outings". Move that section away from festival.yaml into its own data file.

Remove price in the excursions/activities page because it varies depending on program.

Specify which outings are available at the program level instead of at the festival level. Make it possible to override the cost. If it's 0, it's listed as included.