Improve mobile layout
Images that are in sidebar now fill the entire width of the viewport (sans padding).

The additional payment buttons should have some space in them. Avoid adding the ":" suffix to the "Application Form" button.

Make the underline distance slightly shorter. It was competing with the next line, when outings listed the programs that had them as available options.