Switch faculty to people
Since some of the bios are not faculty, change the concept of people participating in the festival as people.

By default (see _config.yml), all individuals in the _people/ directory still have their permalink go to "/faculty/" since I expect most new additions will be faculty. Instructors for other programs, or staff, should have the "permalink" line included in their front matter.

Change the permalinks of the people not affiliated with the piano program to "/people/". Update includes and parameters to refer to "person" or "people" instead of "faculty".