Fix bug where hashtag with two dashes turns to en dash
The use of markdownify-without-p has the side effect of turning a hashtag like "#cancellation-policy--refunds" into "#cancellation-policy–refunds", as in with an en dash.

There isn't a good workaround for this other than to explicitly name the id for that header.

Also, in localize.html, remove a redundant smartify. markdownify-without-p.html already smartifies the text