Fix #51 (sort the faculty tiles on piano front brochure by last name)
Collect people into an array then sort by last name, rather than going through each indiviually and assuming the list is sorted. There is a backup that looks at staff, but the staff doesn't have last name, so it won't work. When we hit that case we will need to refactor.