Fix #46 (link program heads to bios)
Create a new person-localized-name.fx which returns the localized name based on the page language. Refactor faculty-tiles to use this. Wrap the ZH name (if prefers-en-name) in a span, similar to the person bio page. Style appropriately as a block in faculty tiles, but as an inline span for the program heads section of a program page.