Number of localization fixed (complete fix #48)
Application sidebar removes whitespace in between elements so that ZH can have all the text abut each other. Style it so it is the same size for ZH.

Remove extra space after the cost in cancellation table for ZH.

Update the language of the page in <html>. This has the side effect of changing default header fonts.

Define more brand font weights. Use the heavier weight so that the Chinese and Western characters have similar font stroke widths.

Use different colons depends on lang.

Fly-by simplified use of :where() rule.