Fix bug where programs are cut off
If you clicked on a program link in the header or footer, and it was TBA, you go to the home page, and it takes you to #program which is the header for the festival listing. That whole panel would be cut off at the top.

Have it go to #festival instead, which contains that section.